+++ lib/cask/artifact/uninstall.rb
@@ -1,0 +1,2 @@
+class Cask::Artifact::Uninstall < Cask::Artifact::UninstallBase
+end
